Implemented TSK systems. The usage is like: tip = fcontrol.Consequent(np.array([0, 100]), 'tip') #0 and 100 are                     #the minimum and maximum of the variable domain tip['increasing'] = fuzz.Polynomial(domain=tip.universe, expression="5*quality+5*service")
TSK consequents are used as Fuzzy terms. The difference is that in TSK
terms the mf attribute is a Polynomial object, which contains the domain
and the expression (polynomial), which is a string using the same names of the
input variables.

However, this PR may still require some modifications (change some names, among
other issues).