Incorporates out of range values
Currently, a value out of the range of the set will throw an error as line x[x <= xx][-1] in the interp_membership function returns an empty list. (The value being compared is not in the range of the 1D array denoted by x)

This particular change will circumvent that issue and return 0.0 indicating that the membership is 0 for that set.